+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## 1.3.0 - 2019-10-29
+### Changed
+- `composer.json` updated for Laravel 5.7  
+- Bugfix in `Gecche\Multidomain\Queue\Listener` due to changes in handling 
+worker commands in the parent class.
+
 ## 1.2.0 - 2019-10-29
 ### Changed
 - `composer.json` updated for Laravel 5.6  
@@ -53,3 +59,4 @@ Laravel release starting from 5.5 and as pointed out in the docs)
 [1.1.5]: https://github.com/gecche/laravel-multidomain/compare/v1.1.4...v1.1.5
 [1.1.6]: https://github.com/gecche/laravel-multidomain/compare/v1.1.5...v1.1.6
 [1.2.0]: https://github.com/gecche/laravel-multidomain/compare/v1.1.6...v1.2.0
+[1.3.0]: https://github.com/gecche/laravel-multidomain/compare/v1.2.0...v1.3.0